Clean up code and fix bug during split hands
I refactored blackjack, to be a bit more clean with some helper
functions. I also fixed a bug where if you bust in your second hand
during a split play, it counts as an automatic loss for both hands. I
also added some indicators of which hand you are currently playing, and
the winnings for each hand.